+++ CHANGELOG.md
@@ -6,7 +6,12 @@ Full documentation for for rocPyDecode is available at [https://rocm.docs.amd.co
 
 ### Added
 
-* rocpyjpegdecode package 
+* rocpyjpegdecode package
+* Added src/rocjpeg source new subfolder
+* Moved rocPyDecode source files into new subfolder src/rocdecode
+* Moved the dlpack and buffer CPP & H files into src/common new subfolder
+* Created new samples/rocdecode sub-folder
+* Moved rocPyDecode Python Samples under samples/rocdecode
 
 ### Changed